Rename all appearances of io_fleurxml
