+++ masci_tools/cmdline/commands/fleur_schema.py
@@ -8,7 +8,7 @@ import masci_tools
 from masci_tools.cmdline.utils import echo
 from masci_tools.util.xml.common_functions import clear_xml
 from masci_tools.util.xml.converters import convert_str_version_number
-from masci_tools.io.io_fleurxml import load_inpxml, load_outxml
+from masci_tools.io.fleur_xml import load_inpxml, load_outxml
 from masci_tools.io.parsers.fleur import inpxml_parser, outxml_parser
 from masci_tools.io.parsers.fleur_schema import InputSchemaDict, OutputSchemaDict, list_available_versions
